+++ src/exceptions/store_exceptions.cpp
@@ -5,9 +5,9 @@
 #include "store_exceptions.h"
 
 StoreDoesNotExist::StoreDoesNotExist(const std::string& name) :
-    std::invalid_argument(name + " store does not exist!"){
+    std::invalid_argument(name + " Store does not exist!"){
 }
 
 StoreAlreadyExists::StoreAlreadyExists(const std::string &name) :
-    logic_error(name + " store already exists!") {
+    logic_error(name + " Store already exists!") {
 }
