+++ src/model/product.cpp
@@ -2,7 +2,7 @@
 // Created by laimi on 15/10/2020.
 //
 
-#include "Product.h"
+#include "product.h"
 
 #include <utility>
