+++ src/model/store.cpp
@@ -2,7 +2,7 @@
 // Created by laimi on 15/10/2020.
 //
 
-#include "Store.h"
+#include "store.h"
 
 #include <algorithm>
